core_engine/requirements.txt
pdfplumber
google-genai
reportlab